requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
twilio==9.7.2
openai==1.102.0
python-dotenv==1.1.1
pinecone==7.3.0
jinja2==3.1.6
python-multipart==0.0.9
guardrails-ai==0.6.6

